column lineage for external table lineage
