+++ ingestion/src/metadata/ingestion/source/database/athena/metadata.py
@@ -12,13 +12,9 @@
 """Athena source module"""
 
 import traceback
-from copy import deepcopy
-from typing import Dict, Iterable, List, Optional, Tuple
+from typing import Iterable, Optional, Tuple
 
 from pyathena.sqlalchemy.base import AthenaDialect
-from pyathena.sqlalchemy.util import _HashableDict
-from sqlalchemy import types
-from sqlalchemy.engine import reflection
 from sqlalchemy.engine.reflection import Inspector
 
 from metadata.generated.schema.entity.data.databaseSchema import DatabaseSchema
@@ -43,9 +39,13 @@ from metadata.ingestion.api.models import Either
 from metadata.ingestion.api.steps import InvalidSourceException
 from metadata.ingestion.models.ometa_classification import OMetaTagAndClassification
 from metadata.ingestion.ometa.ometa_api import OpenMetadata
-from metadata.ingestion.source import sqa_types
 from metadata.ingestion.source.database.athena.client import AthenaLakeFormationClient
-from metadata.ingestion.source.database.column_type_parser import ColumnTypeParser
+from metadata.ingestion.source.database.athena.utils import (
+    _get_column_type,
+    get_columns,
+    get_table_options,
+    get_view_definition,
+)
 from metadata.ingestion.source.database.common_db_source import (
     CommonDbSourceService,
     TableNameAndType,
@@ -55,12 +55,16 @@ from metadata.ingestion.source.database.external_table_lineage_mixin import (
 )
 from metadata.utils import fqn
 from metadata.utils.logger import ingestion_logger
-from metadata.utils.sqlalchemy_utils import (
-    get_all_table_ddls,
-    get_table_ddl,
-    is_complex_type,
-)
+from metadata.utils.sqlalchemy_utils import get_all_table_ddls, get_table_ddl
 from metadata.utils.tag_utils import get_ometa_tag_and_classification
+
+AthenaDialect._get_column_type = _get_column_type  # pylint: disable=protected-access
+AthenaDialect.get_columns = get_columns
+AthenaDialect.get_view_definition = get_view_definition
+AthenaDialect.get_table_options = get_table_options
+
+Inspector.get_all_table_ddls = get_all_table_ddls
+Inspector.get_table_ddl = get_table_ddl
 
 logger = ingestion_logger()
 
@@ -78,183 +82,6 @@ ATHENA_INTERVAL_TYPE_MAP = {
     ),
     "injected": PartitionIntervalTypes.INJECTED,
 }
-
-
-def _get_column_type(self, type_):
-    """
-    Function overwritten from AthenaDialect
-    to add custom SQA typing.
-    """
-    type_ = type_.replace(" ", "").lower()
-    match = self._pattern_column_type.match(type_)  # pylint: disable=protected-access
-    if match:
-        name = match.group(1).lower()
-        length = match.group(2)
-    else:
-        name = type_.lower()
-        length = None
-
-    args = []
-    col_map = {
-        "boolean": types.BOOLEAN,
-        "float": types.FLOAT,
-        "double": types.FLOAT,
-        "real": types.FLOAT,
-        "tinyint": types.INTEGER,
-        "smallint": types.INTEGER,
-        "integer": types.INTEGER,
-        "int": types.INTEGER,
-        "bigint": types.BIGINT,
-        "string": types.String,
-        "date": types.DATE,
-        "timestamp": types.TIMESTAMP,
-        "binary": types.BINARY,
-        "varbinary": types.BINARY,
-        "array": types.ARRAY,
-        "json": types.JSON,
-        "struct": sqa_types.SQAStruct,
-        "row": sqa_types.SQAStruct,
-        "map": sqa_types.SQAMap,
-        "decimal": types.DECIMAL,
-        "varchar": types.VARCHAR,
-        "char": types.CHAR,
-    }
-    if name in ["decimal", "char", "varchar"]:
-        col_type = col_map[name]
-        if length:
-            args = [int(l) for l in length.split(",")]
-    elif type_.startswith("array"):
-        parsed_type = (
-            ColumnTypeParser._parse_datatype_string(  # pylint: disable=protected-access
-                type_
-            )
-        )
-        col_type = col_map["array"]
-        if parsed_type["arrayDataType"].lower().startswith("array"):
-            # as OpenMetadata doesn't store any details on children of array, we put
-            # in type as string as default to avoid Array item_type required issue
-            # from sqlalchemy types
-            args = [types.String]
-        else:
-            args = [col_map.get(parsed_type.get("arrayDataType").lower(), types.String)]
-    elif col_map.get(name):
-        col_type = col_map.get(name)
-    else:
-        logger.warning(f"Did not recognize type '{type_}'")
-        col_type = types.NullType
-    return col_type(*args)
-
-
-def _get_projection_details(
-    columns: List[Dict], projection_parameters: Dict
-) -> List[Dict]:
-    """Get the projection details for the columns
-
-    Args:
-        columns (List[Dict]): list of columns
-        projection_parameters (Dict): projection parameters
-    """
-    if not projection_parameters:
-        return columns
-
-    columns = deepcopy(columns)
-    for col in columns:
-        projection_details = next(
-            ({k: v} for k, v in projection_parameters.items() if k == col["name"]), None
-        )
-        if projection_details:
-            col["projection_type"] = projection_details[col["name"]]
-
-    return columns
-
-
-@reflection.cache
-def get_columns(self, connection, table_name, schema=None, **kw):
-    """
-    Method to handle table columns
-    """
-    metadata = self._get_table(  # pylint: disable=protected-access
-        connection, table_name, schema=schema, **kw
-    )
-    columns = [
-        {
-            "name": c.name,
-            "type": self._get_column_type(c.type),  # pylint: disable=protected-access
-            "nullable": True,
-            "default": None,
-            "autoincrement": False,
-            "comment": c.comment,
-            "system_data_type": c.type,
-            "is_complex": is_complex_type(c.type),
-            "dialect_options": {"awsathena_partition": True},
-        }
-        for c in metadata.partition_keys
-    ]
-
-    if kw.get("only_partition_columns"):
-        # Return projected partition information to set partition type in `get_table_partition_details`
-        # projected partition fields are stored in the form of `projection.<field_name>.type` as a table parameter
-        projection_parameters = {
-            key_.split(".")[1]: value_
-            for key_, value_ in metadata.parameters.items()
-            if key_.startswith("projection") and key_.endswith("type")
-        }
-        columns = _get_projection_details(columns, projection_parameters)
-        return columns
-
-    columns += [
-        {
-            "name": c.name,
-            "type": self._get_column_type(c.type),  # pylint: disable=protected-access
-            "nullable": True,
-            "default": None,
-            "autoincrement": False,
-            "comment": c.comment,
-            "system_data_type": c.type,
-            "is_complex": is_complex_type(c.type),
-            "dialect_options": {"awsathena_partition": None},
-        }
-        for c in metadata.columns
-    ]
-
-    return columns
-
-
-# pylint: disable=unused-argument
-@reflection.cache
-def get_view_definition(self, connection, view_name, schema=None, **kw):
-    """
-    Gets the view definition
-    """
-    full_view_name = f'"{view_name}"' if not schema else f'"{schema}"."{view_name}"'
-    res = connection.execute(f"SHOW CREATE VIEW {full_view_name}").fetchall()
-    if res:
-        return "\n".join(i[0] for i in res)
-    return None
-
-
-@reflection.cache
-def get_table_options(
-    self, connection: "Connection", table_name: str, schema: Optional[str] = None, **kw
-):
-    metadata = self._get_table(connection, table_name, schema=schema, **kw)
-    # TODO The metadata retrieved from the API does not seem to include bucketing information.
-    return {
-        "awsathena_location": metadata.location,
-        "awsathena_compression": metadata.compression,
-        "awsathena_row_format": metadata.row_format,
-        "awsathena_file_format": metadata.file_format,
-        "awsathena_serdeproperties": _HashableDict(metadata.serde_properties),
-        "awsathena_tblproperties": _HashableDict(metadata.table_properties),
-    }
-
-
-AthenaDialect._get_column_type = _get_column_type  # pylint: disable=protected-access
-AthenaDialect.get_columns = get_columns
-AthenaDialect.get_view_definition = get_view_definition
-AthenaDialect.get_table_options = get_table_options
-Inspector.get_all_table_ddls = get_all_table_ddls
-Inspector.get_table_ddl = get_table_ddl
 
 
 class AthenaSource(ExternalTableLineageMixin, CommonDbSourceService):
