+++ ingestion/src/metadata/ingestion/source/database/external_table_lineage_mixin.py
@@ -14,13 +14,20 @@ External Table Lineage Mixin
 
 import traceback
 from abc import ABC
-from typing import Iterable
+from typing import Iterable, List, Optional
 
 from metadata.generated.schema.api.lineage.addLineage import AddLineageRequest
+from metadata.generated.schema.entity.data.container import ContainerDataModel
 from metadata.generated.schema.entity.data.table import Table
-from metadata.generated.schema.type.entityLineage import EntitiesEdge
+from metadata.generated.schema.type.entityLineage import (
+    ColumnLineage,
+    EntitiesEdge,
+    LineageDetails,
+)
+from metadata.generated.schema.type.entityLineage import Source as LineageSource
 from metadata.generated.schema.type.entityReference import EntityReference
 from metadata.ingestion.api.models import Either
+from metadata.ingestion.lineage.sql_lineage import get_column_fqn
 from metadata.utils import fqn
 from metadata.utils.logger import ingestion_logger
 
@@ -39,7 +46,7 @@ class ExternalTableLineageMixin(ABC):
         for table_qualified_tuple, location in self.external_location_map.items() or []:
             try:
                 location_entity = self.metadata.es_search_container_by_path(
-                    full_path=location
+                    full_path=location, fields="dataModel"
                 )
                 database_name, schema_name, table_name = table_qualified_tuple
 
@@ -63,6 +70,12 @@ class ExternalTableLineageMixin(ABC):
                     and table_entity
                     and table_entity[0]
                 ):
+                    columns_list = [
+                        column.name.__root__ for column in table_entity[0].columns
+                    ]
+                    columns_lineage = self._get_column_lineage(
+                        location_entity[0].dataModel, table_entity[0], columns_list
+                    )
                     yield Either(
                         right=AddLineageRequest(
                             edge=EntitiesEdge(
@@ -74,9 +87,51 @@ class ExternalTableLineageMixin(ABC):
                                     id=table_entity[0].id,
                                     type="table",
                                 ),
+                                lineageDetails=LineageDetails(
+                                    source=LineageSource.ExternalTableLineage,
+                                    columnsLineage=columns_lineage,
+                                ),
                             )
                         )
                     )
             except Exception as exc:
                 logger.warning(f"Failed to yield external table lineage due to - {exc}")
                 logger.debug(traceback.format_exc())
+
+    def _get_data_model_column_fqn(
+        self, data_model_entity: ContainerDataModel, column: str
+    ) -> Optional[str]:
+        """
+        Get fqn of column if exist in data model entity
+        """
+        if not data_model_entity:
+            return None
+        for entity_column in data_model_entity.columns:
+            if entity_column.displayName.lower() == column.lower():
+                return entity_column.fullyQualifiedName.__root__
+        return None
+
+    def _get_column_lineage(
+        self,
+        data_model_entity: ContainerDataModel,
+        table_entity: Table,
+        columns_list: List[str],
+    ) -> List[ColumnLineage]:
+        """
+        Get the column lineage
+        """
+        try:
+            column_lineage = []
+            for field in columns_list or []:
+                from_column = self._get_data_model_column_fqn(
+                    data_model_entity=data_model_entity, column=field
+                )
+                to_column = get_column_fqn(table_entity=table_entity, column=field)
+                if from_column and to_column:
+                    column_lineage.append(
+                        ColumnLineage(fromColumns=[from_column], toColumn=to_column)
+                    )
+            return column_lineage
+        except Exception as exc:
+            logger.debug(f"Error to get column lineage: {exc}")
+            logger.debug(traceback.format_exc())
